+++ packages/atomic/src/Link/index.tsx
@@ -9,13 +9,12 @@ export type LinkProps = BaseProps<React.LinkHTMLAttributes<HTMLLinkElement>> & {
 };
 
 export const Link: React.FC<LinkProps> = (props: LinkProps) => {
-  const { css, className, to, ...restProps } = props;
+  const { css = {}, className, to, ...restProps } = props;
   return (
-    //@ts-ignore
     <Atomic
       as="a"
       className={classNames(className, 'c3-link')}
-      css={{ ...link(), ...css }}
+      css={{ ...css }}
       href={to}
       {...restProps}
     />