+++ packages/atomic/src/Text/text.utils.ts
@@ -1,17 +1,19 @@
 import type { Typography } from './types';
-import type { CSSProperties } from '@unstyled-ui/css';
+import type { CSSProperties } from '@unstyled-ui/core';
 
 export const linearGradientText = (
   bg: CSSProperties['background']
 ): CSSProperties => ({
   background: bg,
   backgroundClip: 'text',
+  //@ts-ignore
   WebkitBackgroundClip: 'text',
   WebkitTextFillColor: 'transparent',
 });
 
 export const getRows = (rows: number): CSSProperties => ({
   display: '-webkit-box',
+  //@ts-ignore
   WebkitLineClamp: rows,
   WebkitBoxOrient: 'vertical',
   overflow: 'hidden',