+++ packages/core/src/types.ts
@@ -2,6 +2,7 @@ import type { CSS } from '@unstyled-ui/stitches/types/css-util.d';
 import type { Config } from './configure';
 import type { VariantProps } from '@unstyled-ui/stitches';
 import type { CSSProperties } from '@unstyled-ui/stitches';
+export type { CSSProperties } from '@unstyled-ui/stitches';
 
 export type RCSSProperties = {
   [k in keyof CSSProperties]?: CSSProperties[k] | CSSProperties[k][];
@@ -17,10 +18,8 @@ export type RCSSProps = {
   [k in keyof CSSProps]?: CSSProps[k] | CSSProps[k][];
 };
 
-
-
 export type BaseProps<
   Attr extends React.HTMLAttributes<HTMLElement> = React.HTMLAttributes<HTMLElement>
 > = VariantProps<any> & {
-  css?: CSSProps;
+  css?: RCSSProps;
 } & Attr;