+++ packages/css/src/utils/mask.ts
@@ -1,4 +1,4 @@
-import { CSSProperties } from './../types';
+import { CSSProperties } from '@unstyled-ui/core';
 
 export const mask: CSSProperties = {
   position: 'fixed',