+++ packages/charts/CHANGELOG.md
@@ -1,1 +1,9 @@
+## 1.0.2
 
+### Patch Changes
+
+- Updated dependencies [be9cc34]
+- Updated dependencies [b0f317f]
+  - @unstyled-ui/css@1.0.2
+  - @unstyled-ui/stitches@1.0.4
+  - @unstyled-ui/core@1.0.2