+++ packages/stitches/CHANGELOG.md
@@ -1,5 +1,16 @@
 ## 1.0.2
 
+## 1.0.4
+
+### Patch Changes
+
+- be9cc34: add ui-basic package
+- b0f317f: add storybook support
+- Updated dependencies [be9cc34]
+- Updated dependencies [b0f317f]
+  - @unstyled-ui/css@1.0.2
+  - @unstyled-ui/responsive@1.0.2
+
 ### Patch Changes
 
 - 8a34fb6: fix resolve @unstyled-ui/cli error